Enable dataset validation
