+++ util/dataset.go
@@ -85,7 +85,6 @@ type Os struct {
 }
 
 func NewDataset(basePath string, p *Package) (*DataSet, error) {
-
 	// Check if manifest exists
 	manifestPath := filepath.Join(basePath, "manifest.yml")
 	_, err := os.Stat(manifestPath)
@@ -132,7 +131,6 @@ func NewDataset(basePath string, p *Package) (*DataSet, error) {
 	if !IsValidRelease(d.Release) {
 		return nil, fmt.Errorf("invalid release: %s", d.Release)
 	}
-
 	return d, nil
 }
 
@@ -141,10 +139,6 @@ func (d *DataSet) Validate() error {
 	paths, err := filepath.Glob(filepath.Join(pipelineDir, "*"))
 	if err != nil {
 		return err
-	}
-
-	if strings.Contains(d.ID, "-") {
-		return fmt.Errorf("dataset name is not allowed to contain `-`: %s", d.ID)
 	}
 
 	if !d.validType() {
@@ -199,6 +193,11 @@ func (d *DataSet) Validate() error {
 			return fmt.Errorf("defined ingest_pipeline does not exist: %s", pipelineDir+d.IngestPipeline)
 		}
 	}
+
+	err = d.validateRequiredFields()
+	if err != nil {
+		return errors.Wrap(err, "validating required fields failed")
+	}
 	return nil
 }
 
@@ -229,4 +228,99 @@ func validateIngestPipelineFile(pipelinePath string) error {
 		return fmt.Errorf("unsupported pipeline extension (path: %s, ext: %s)", pipelinePath, ext)
 	}
 	return err
+}
+
+// validateRequiredFields method loads fields from all files and checks if required fields are present.
+func (d *DataSet) validateRequiredFields() error {
+	fieldsDirPath := filepath.Join(d.BasePath, "fields")
+
+	// Collect fields from all files
+	var allFields []MapStr
+	err := filepath.Walk(fieldsDirPath, func(path string, info os.FileInfo, err error) error {
+		if err != nil {
+			return err
+		}
+
+		relativePath, err := filepath.Rel(fieldsDirPath, path)
+		if err != nil {
+			return errors.Wrapf(err, "cannot find relative path (fieldsDirPath: %s, path: %s)", fieldsDirPath, path)
+		}
+
+		if relativePath == "." {
+			return nil
+		}
+
+		body, err := ioutil.ReadFile(path)
+		if err != nil {
+			return errors.Wrapf(err, "reading file failed (path: %s)", path)
+		}
+
+		var m []MapStr
+		err = yamlv2.Unmarshal(body, &m)
+		if err != nil {
+			return errors.Wrapf(err, "unmarshaling file failed (path: %s)", path)
+		}
+
+		allFields = append(allFields, m...)
+		return nil
+	})
+	if err != nil {
+		return errors.Wrapf(err, "walking through fields files failed")
+	}
+
+	// Flatten all fields
+	for i, fields := range allFields {
+		allFields[i] = fields.Flatten()
+	}
+
+	// Verify required keys
+	err = requireField(allFields, "dataset.type", "constant_keyword", err)
+	err = requireField(allFields, "dataset.name", "constant_keyword", err)
+	err = requireField(allFields, "dataset.namespace", "constant_keyword", err)
+	err = requireField(allFields, "@timestamp", "date", err)
+	return err
+}
+
+func requireField(allFields []MapStr, searchedName, expectedType string, validationErr error) error {
+	if validationErr != nil {
+		return validationErr
+	}
+
+	f, err := findField(allFields, searchedName)
+	if err != nil {
+		return errors.Wrapf(err, "finding field failed (searchedName: %s)", searchedName)
+	}
+
+	if f.aType != expectedType {
+		return fmt.Errorf("wrong field type for '%s' (expected: %s, got: %s)", searchedName, expectedType, f.aType)
+	}
+	return nil
+}
+
+func findField(allFields []MapStr, searchedName string) (*fieldEntry, error) {
+	for _, fields := range allFields {
+		name, err := fields.GetValue("name")
+		if err != nil {
+			return nil, errors.Wrapf(err, "cannot get value (key: name)")
+		}
+
+		if name != searchedName {
+			continue
+		}
+
+		aType, err := fields.GetValue("type")
+		if err != nil {
+			return nil, errors.Wrapf(err, "cannot get value (key: type)")
+		}
+
+		if aType == "" {
+			return nil, fmt.Errorf("field '%s' found, but type is undefined", searchedName)
+		}
+
+		return &fieldEntry{
+			name:  name.(string),
+			aType: aType.(string),
+		}, nil
+	}
+	return nil, fmt.Errorf("field '%s' not found", searchedName)
 }
